Refactor DateRange/DateSelect
