+++ src/formBuilder/dateRange.jsx
@@ -1,47 +1,34 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import { Flex } from '@chakra-ui/react';
-import DateSelect from '../components/dateSelect';
+import DateSelect from './dateSelect';
 import FieldWrapper from './fieldWrapper';
 
-const DateRange = ({ maxDays }) => {
+const DateRange = props => {
+  const { localForm } = props;
+  const { watch } = localForm;
   const today = new Date();
-  const [startDate, setStartDate] = useState(null);
-  const [endDate, setEndDate] = useState();
-  const [maxDate, setMaxDate] = useState();
-
-  useEffect(() => {
-    if (!startDate) {
-      setMaxDate(null);
-    } else {
-      const newDate = new Date(startDate);
-      newDate.setDate(newDate.getDate() + maxDays);
-      setMaxDate(newDate);
-    }
-  }, [startDate]);
+  const startDate = watch('startDate');
 
   return (
     <FieldWrapper>
       <Flex flexDir='row' justify='space-between'>
         <DateSelect
+          {...props}
+          htmlFor='startDate'
+          name='startDate'
           width='48%'
           placeholderText='Start DateTime'
-          showTimeInput
-          selected={startDate}
-          onChange={date => {
-            setStartDate(date);
-          }}
+          showTimeSelect
           minDate={today}
         />
         <DateSelect
+          {...props}
+          htmlFor='endDate'
+          name='endDate'
           width='48%'
           placeholderText='End DateTime'
-          showTimeInput
-          selected={endDate}
-          onChange={date => {
-            setEndDate(date);
-          }}
-          minDate={startDate}
-          maxDate={maxDate}
+          showTimeSelect
+          minDate={new Date(startDate)}
         />
       </Flex>
     </FieldWrapper>
